+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add header, breadcrumb, prev/next nav, tags, RSS link, and SEO improvements
+
 ### Changed
 
 - Remove roadmap app documentation